// src/app/model/InfoCardVO.ts
/**
 * Wision55
 * API per l'utilizzo dei servizi della piattaforma Wision55
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface InfoCardVO {
    id?: number;

    saldo?: number;

    attivo?: boolean;

    nome?: string;

    cognome?: string;

    comune?: string;
    
    cliente_id?: number;

    provincia?: string;

    eta?: string;

}