// src/app/model/RispostaOperazioneVO.ts
/**
 * Wision55
 * API per l'utilizzo dei servizi della piattaforma Wision55
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface RispostaOperazioneVO {
    codice_operazione?: number;

    cashback?: number;

    meccanismo_protezione?: RispostaOperazioneVO.MeccanismoProtezioneEnum;

    cifre_pin_parziale?: Array<number>;

    firma_hotp?: string;

}
export namespace RispostaOperazioneVO {
    export enum MeccanismoProtezioneEnum {
        PinTotale = <any> 'pin_totale',
        PinParziale = <any> 'pin_parziale',
        TOtp = <any> 't_otp',
        HOtp = <any> 'h_otp',
        CodiceSms = <any> 'codice_sms',
        Nessuno = <any> 'nessuno'
    }
}