// src/app/model/TransazioneVO.ts
/**
 * Wision55
 * API per l'utilizzo dei servizi della piattaforma Wision55
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface TransazioneVO {
    id?: number;

    tipo?: string;

    token?: string;

    da_chi?: string;

    a_chi?: string;

    descrizione?: string;

    cashbackWKey?: number;

    totaleOperazione?: number;

    cashback?: number;

    prelievoCarta?: number;

    negozioId?: number;

    posId?: number;

    cardId?: number;

    cliente_id?: number;

    stato?: string;

    nominativo_cliente?: string;

    data_richiesta?: Date;

    data_completamento?: Date;

    dataAnnullamento?: Date;

}